deploy/pyproject.toml
[project]
name = "openclaw-orchestrator"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "aws-cdk-lib>=2.170.0",
    "constructs>=10.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest", "boto3-stubs"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = []
